A task to cleanup tekton objects.
This is a task that wrap a python script to let to schedule a cleanup of old
tekton objects (usually runs).

It can works via age (keep the last 5 days) or via a limit number (keep the last
5 pipelineruns) and if you have the rights let you pass a target namespace or do
some filtering by labels.

This task can be run via kubernetes or 'vixie' style cron.